Deduplicate root derive code
