+++ binrw_derive/src/lib.rs
@@ -10,46 +10,67 @@
 
 #[cfg(all(nightly, not(coverage)))]
 mod backtrace;
-mod binread;
 mod binrw_attr;
-mod binwrite;
 mod codegen;
 mod named_args;
 mod parser;
 
-use crate::{
-    codegen::typed_builder::{Builder, BuilderField, BuilderFieldKind},
-    named_args::NamedArgAttr,
-};
+use codegen::{generate_binread_impl, generate_binwrite_impl};
+use parser::{Input, ParseResult};
 use proc_macro::TokenStream;
-use syn::{parse_macro_input, spanned::Spanned, DeriveInput};
+use quote::quote;
+use syn::{parse_macro_input, DeriveInput};
 
 #[proc_macro_derive(BinRead, attributes(binread, br, brw))]
 #[cfg_attr(coverage_nightly, no_coverage)]
 pub fn derive_binread_trait(input: TokenStream) -> TokenStream {
-    binread::derive_from_input(&parse_macro_input!(input as DeriveInput), true)
-        .1
-        .into()
+    derive_from_input(
+        parse_macro_input!(input as DeriveInput),
+        Options {
+            derive: true,
+            write: false,
+        },
+    )
+    .into()
 }
 
 #[proc_macro_attribute]
 #[cfg_attr(coverage_nightly, no_coverage)]
 pub fn binread(_: TokenStream, input: TokenStream) -> TokenStream {
-    binread::derive_from_attribute(parse_macro_input!(input as DeriveInput)).into()
+    derive_from_input(
+        parse_macro_input!(input as DeriveInput),
+        Options {
+            derive: false,
+            write: false,
+        },
+    )
+    .into()
 }
 
 #[proc_macro_derive(BinWrite, attributes(binwrite, bw, brw))]
 #[cfg_attr(coverage_nightly, no_coverage)]
 pub fn derive_binwrite_trait(input: TokenStream) -> TokenStream {
-    binwrite::derive_from_input(&parse_macro_input!(input as DeriveInput), true)
-        .1
-        .into()
+    derive_from_input(
+        parse_macro_input!(input as DeriveInput),
+        Options {
+            derive: true,
+            write: true,
+        },
+    )
+    .into()
 }
 
 #[proc_macro_attribute]
 #[cfg_attr(coverage_nightly, no_coverage)]
 pub fn binwrite(_: TokenStream, input: TokenStream) -> TokenStream {
-    binwrite::derive_from_attribute(parse_macro_input!(input as DeriveInput)).into()
+    derive_from_input(
+        parse_macro_input!(input as DeriveInput),
+        Options {
+            derive: false,
+            write: true,
+        },
+    )
+    .into()
 }
 
 #[proc_macro_attribute]
@@ -58,80 +79,127 @@ pub fn binrw(_: TokenStream, input: TokenStream) -> TokenStream {
     binrw_attr::derive_from_attribute(parse_macro_input!(input as DeriveInput)).into()
 }
 
-fn binrw_named_args(input: DeriveInput) -> proc_macro2::TokenStream {
-    let fields = match match input.data {
-        syn::Data::Struct(s) => s
-            .fields
-            .iter()
-            .map(|field| {
-                let attrs: Vec<NamedArgAttr> = field
-                    .attrs
-                    .iter()
-                    .filter_map(|attr| {
-                        let is_named_args = attr
-                            .path
-                            .get_ident()
-                            .map_or(false, |ident| ident == "named_args");
-                        if is_named_args {
-                            attr.parse_args().ok()
-                        } else {
-                            None
-                        }
-                    })
-                    .collect();
-                let kind = if attrs
-                    .iter()
-                    .any(|attr| matches!(attr, NamedArgAttr::TryOptional))
-                {
-                    BuilderFieldKind::TryOptional
-                } else if let Some(NamedArgAttr::Default(default)) = attrs
-                    .iter()
-                    .find(|attr| matches!(attr, NamedArgAttr::Default(_)))
-                {
-                    BuilderFieldKind::Optional {
-                        default: default.clone(),
-                    }
-                } else {
-                    BuilderFieldKind::Required
-                };
-                Ok(BuilderField {
-                    kind,
-                    name: match field.ident.as_ref() {
-                        Some(ident) => ident.clone(),
-                        None => {
-                            return Err(syn::Error::new(
-                                field.span(),
-                                "must not be a tuple-style field",
-                            ))
-                        }
-                    },
-                    ty: field.ty.clone(),
-                })
-            })
-            .collect::<Result<Vec<_>, syn::Error>>(),
-        _ => return syn::Error::new(input.span(), "only structs are supported").to_compile_error(),
-    } {
-        Ok(fields) => fields,
-        Err(err) => return err.into_compile_error(),
-    };
+#[proc_macro_derive(BinrwNamedArgs, attributes(named_args))]
+#[cfg_attr(coverage_nightly, no_coverage)]
+pub fn derive_binrw_named_args(input: TokenStream) -> TokenStream {
+    named_args::derive_from_attribute(parse_macro_input!(input as DeriveInput)).into()
+}
 
-    let generics: Vec<_> = input.generics.params.iter().cloned().collect();
-
-    Builder {
-        result_name: &input.ident,
-        builder_name: &quote::format_ident!("{}Builder", input.ident),
-        fields: &fields,
-        generics: &generics,
-        vis: &input.vis,
-    }
-    .generate(false)
+/// Input handling options.
+#[derive(Clone, Copy)]
+struct Options {
+    /// If `true`, the input is from a `#[derive]` instead of an attribute.
+    derive: bool,
+    /// If `true`, the input is for `BinWrite`.
+    write: bool,
 }
 
 #[cfg_attr(coverage_nightly, no_coverage)]
-#[proc_macro_derive(BinrwNamedArgs, attributes(named_args))]
-pub fn derive_binrw_named_args(input: TokenStream) -> TokenStream {
-    let input = parse_macro_input!(input as DeriveInput);
-    binrw_named_args(input).into()
+fn clean_field_attrs(input: &Option<Input>, variant_index: usize, fields: &mut syn::Fields) {
+    if let Some(input) = input {
+        let fields = match fields {
+            syn::Fields::Named(fields) => &mut fields.named,
+            syn::Fields::Unnamed(fields) => &mut fields.unnamed,
+            syn::Fields::Unit => return,
+        };
+
+        *fields = fields
+            .iter_mut()
+            .enumerate()
+            .filter_map(|(index, value)| {
+                if input.is_temp_field(variant_index, index) {
+                    None
+                } else {
+                    let mut value = value.clone();
+                    clean_struct_attrs(&mut value.attrs);
+                    Some(value)
+                }
+            })
+            .collect();
+    }
+}
+
+#[cfg_attr(coverage_nightly, no_coverage)]
+fn clean_struct_attrs(attrs: &mut Vec<syn::Attribute>) {
+    attrs.retain(|attr| !is_binwrite_attr(attr) && !is_binread_attr(attr));
+}
+
+#[cfg_attr(coverage_nightly, no_coverage)]
+fn derive_from_input(mut derive_input: DeriveInput, options: Options) -> proc_macro2::TokenStream {
+    let (binrw_input, generated_impl) = parse(&derive_input, options);
+    let binrw_input = binrw_input.ok();
+
+    // only clean fields if binwrite isn't going to be applied after
+    if has_attr(
+        &derive_input,
+        if options.write { "binread" } else { "binwrite" },
+    ) {
+        return quote! {
+            compile_error!("cannot combine `#[binread]` and `#[binwrite]`; use `#[binrw]` instead");
+
+            #derive_input
+            #generated_impl
+        };
+    }
+
+    if options.derive {
+        generated_impl
+    } else {
+        clean_struct_attrs(&mut derive_input.attrs);
+
+        match &mut derive_input.data {
+            syn::Data::Struct(st) => {
+                clean_field_attrs(&binrw_input, 0, &mut st.fields);
+            }
+            syn::Data::Enum(en) => {
+                for (index, variant) in en.variants.iter_mut().enumerate() {
+                    clean_struct_attrs(&mut variant.attrs);
+                    clean_field_attrs(&binrw_input, index, &mut variant.fields);
+                }
+            }
+            syn::Data::Union(union) => {
+                for field in union.fields.named.iter_mut() {
+                    clean_struct_attrs(&mut field.attrs);
+                }
+            }
+        }
+
+        quote!(
+            #derive_input
+            #generated_impl
+        )
+    }
+}
+
+// TODO: make this work for `#[binrw::binwrite]` somehow?
+#[cfg_attr(coverage_nightly, no_coverage)]
+fn has_attr(input: &DeriveInput, attr_name: &str) -> bool {
+    input.attrs.iter().any(|attr| {
+        attr.path
+            .get_ident()
+            .map_or(false, |ident| ident == attr_name)
+    })
+}
+
+fn is_binread_attr(attr: &syn::Attribute) -> bool {
+    attr.path.is_ident("br") || attr.path.is_ident("brw")
+}
+
+fn is_binwrite_attr(attr: &syn::Attribute) -> bool {
+    attr.path.is_ident("bw") || attr.path.is_ident("brw")
+}
+
+fn parse(
+    derive_input: &DeriveInput,
+    options: Options,
+) -> (ParseResult<Input>, proc_macro2::TokenStream) {
+    let binrw_input = Input::from_input(derive_input, options);
+    let generated_impl = if options.write {
+        generate_binwrite_impl(derive_input, &binrw_input)
+    } else {
+        generate_binread_impl(derive_input, &binrw_input)
+    };
+    (binrw_input, generated_impl)
 }
 
 #[cfg(test)]
@@ -154,7 +222,14 @@ fn derive_code_coverage_for_tool() {
         if entry.file_type().unwrap().is_file() {
             let file = fs::File::open(entry.path()).unwrap();
             if emulate_derive_expansion_fallible(file, "BinRead", |input| {
-                binread::derive_from_input(&input, true).1
+                parse(
+                    &input,
+                    Options {
+                        derive: true,
+                        write: false,
+                    },
+                )
+                .1
             })
             .is_err()
             {
@@ -184,7 +259,14 @@ fn derive_binwrite_code_coverage_for_tool() {
         if entry.file_type().unwrap().is_file() {
             let file = fs::File::open(entry.path()).unwrap();
             if emulate_derive_expansion_fallible(file, "BinWrite", |input| {
-                binwrite::derive_from_input(&input, true).1
+                parse(
+                    &input,
+                    Options {
+                        derive: true,
+                        write: true,
+                    },
+                )
+                .1
             })
             .is_err()
             {
