+++ binrw_derive/src/named_args/mod.rs
@@ -1,16 +1,13 @@
+use crate::codegen::typed_builder::{Builder, BuilderField, BuilderFieldKind};
 use syn::{
     parse::{Parse, ParseStream},
-    Expr, Token,
+    spanned::Spanned,
+    DeriveInput, Expr, Token,
 };
 
 pub(crate) enum NamedArgAttr {
     Default(Box<Expr>),
     TryOptional,
-}
-
-mod kw {
-    syn::custom_keyword!(default);
-    syn::custom_keyword!(try_optional);
 }
 
 impl Parse for NamedArgAttr {
@@ -28,4 +25,78 @@ impl Parse for NamedArgAttr {
             Err(lookahead.error())
         }
     }
+}
+
+pub(super) fn derive_from_attribute(input: DeriveInput) -> proc_macro2::TokenStream {
+    let fields = match match input.data {
+        syn::Data::Struct(s) => s
+            .fields
+            .iter()
+            .map(|field| {
+                let attrs: Vec<NamedArgAttr> = field
+                    .attrs
+                    .iter()
+                    .filter_map(|attr| {
+                        let is_named_args = attr
+                            .path
+                            .get_ident()
+                            .map_or(false, |ident| ident == "named_args");
+                        if is_named_args {
+                            attr.parse_args().ok()
+                        } else {
+                            None
+                        }
+                    })
+                    .collect();
+                let kind = if attrs
+                    .iter()
+                    .any(|attr| matches!(attr, NamedArgAttr::TryOptional))
+                {
+                    BuilderFieldKind::TryOptional
+                } else if let Some(NamedArgAttr::Default(default)) = attrs
+                    .iter()
+                    .find(|attr| matches!(attr, NamedArgAttr::Default(_)))
+                {
+                    BuilderFieldKind::Optional {
+                        default: default.clone(),
+                    }
+                } else {
+                    BuilderFieldKind::Required
+                };
+                Ok(BuilderField {
+                    kind,
+                    name: match field.ident.as_ref() {
+                        Some(ident) => ident.clone(),
+                        None => {
+                            return Err(syn::Error::new(
+                                field.span(),
+                                "must not be a tuple-style field",
+                            ))
+                        }
+                    },
+                    ty: field.ty.clone(),
+                })
+            })
+            .collect::<Result<Vec<_>, syn::Error>>(),
+        _ => return syn::Error::new(input.span(), "only structs are supported").to_compile_error(),
+    } {
+        Ok(fields) => fields,
+        Err(err) => return err.into_compile_error(),
+    };
+
+    let generics: Vec<_> = input.generics.params.iter().cloned().collect();
+
+    Builder {
+        result_name: &input.ident,
+        builder_name: &quote::format_ident!("{}Builder", input.ident),
+        fields: &fields,
+        generics: &generics,
+        vis: &input.vis,
+    }
+    .generate(false)
+}
+
+mod kw {
+    syn::custom_keyword!(default);
+    syn::custom_keyword!(try_optional);
 }
